requirements.txt
docx2txt==0.8
gensim==4.3.2
hanlp==2.1.0b52
matplotlib==3.7.1
networkx==3.0
numpy==1.23.5
pandas==1.5.3
python_docx==1.1.0
scikit_learn==1.3.2
seaborn==0.13.0